# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"AccessControlResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"hasMore": {
						"type": "boolean"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateRoleBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccessControlResponse_void_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"nextCursor": {
						"type": "string",
						"nullable": true
					},
					"hasMore": {
						"type": "boolean"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssignRoleBody": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"roleIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"roleIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PermissionInput": {
				"properties": {
					"resourceName": {
						"type": "string"
					},
					"actions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"resourceName",
					"actions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssignPermissionsBody": {
				"properties": {
					"roleId": {
						"type": "string"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/PermissionInput"
						},
						"type": "array"
					}
				},
				"required": [
					"roleId",
					"permissions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "stock-backend",
		"version": "1.0.0",
		"description": "Stock prediction, decision making and analytics platform.",
		"contact": {
			"name": "Hamza Ahmed"
		}
	},
	"paths": {
...
		"/api/v1/rbac/user-screens": {
			"get": {
				"operationId": "GetUserScreenPermissions",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_any_"
								}
							}
						}
					}
				},
				"description": "Get effective screen permissions for the authenticated caller.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/rbac/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_any_"
								}
							}
						}
					}
				},
				"description": "Get all users with paginated list (Admin/RBAC read).",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:READ"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/rbac/roles": {
			"get": {
				"operationId": "GetAllRoles",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_any_"
								}
							}
						}
					}
				},
				"description": "Get all roles defined in the system.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:READ"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "AddRole",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_any_"
								}
							}
						}
					}
				},
				"description": "Create a new role in the system.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:CREATE"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoleBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/rbac/roles/{roleId}": {
			"delete": {
				"operationId": "RevokeRole",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_void_"
								}
							}
						}
					}
				},
				"description": "Revoke a role from a user.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:DELETE"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/rbac/assign-role": {
			"post": {
				"operationId": "AssignRole",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_any_"
								}
							}
						}
					}
				},
				"description": "Assign roles to a user.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:CREATE"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignRoleBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/rbac/permissions": {
			"get": {
				"operationId": "GetAllPermissions",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_any_"
								}
							}
						}
					}
				},
				"description": "Get all permission definitions.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:READ"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/v1/rbac/assign-permissions": {
			"post": {
				"operationId": "AssignPermissionsToRole",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_any_"
								}
							}
						}
					}
				},
				"description": "Assign permissions to a specific role.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:CREATE"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignPermissionsBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/rbac/revoke-permissions": {
			"delete": {
				"operationId": "RevokePermissionsFromRole",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_void_"
								}
							}
						}
					}
				},
				"description": "Revoke permissions from a specific role.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:DELETE"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignPermissionsBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/rbac/resources": {
			"get": {
				"operationId": "GetAllResources",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AccessControlResponse_any_"
								}
							}
						}
					}
				},
				"description": "Get all registered system resources.",
				"tags": [
					"Access Control (RBAC)"
				],
				"security": [
					{
						"bearerAuth": [
							"ROLE:READ"
						]
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}